FlowOps/gateway/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pika==1.3.1 # for RabbitMQ
requests==2.31.0      
python-dotenv==1.0.0  
pydantic==2.4.2
pydantic-settings==2.0.3
debugpy==1.8.0